Adapt to comparisonKey() type change

+++ src/imports/texture-sharing/sharedtextureprovider.h
@@ -100,7 +100,7 @@ public:
     SharedTexture(QtWaylandClient::QWaylandServerBuffer *buffer);
 
     int textureId() const override;
-    int comparisonKey() const override;
+    qint64 comparisonKey() const override;
     QSize textureSize() const override;
     bool hasAlphaChannel() const override;
     bool hasMipmaps() const override;
